plugins/cilium-cni: [2/2] Move implementation into separate package
This commit makes `plugins/cilium-cni/cmd` a proper package and moves
the `main` entry point again back to it's original location. This will
make it easier to re-use parts of the Cilium CNI plugin. Other than
that, this commit contains no functional changes.
